// ProcMgt/ClientApp/src/app/models/itemtype.ts
export class Itemtype {
  itemTypeID: string;
  itemTypeName: string;
  categoryID: string;
  itemTypeCode: string;
  depreciationRate: number;
  measurementUnitID: string;
  approvalPatternTypeID: string;
  isDisposable: boolean;
  isActive: boolean;
  userID: string;
  entryDateTime: string;

  categoryName: string; 
  measurementUnitName: string;
  patternName: string;

} 